requirements.txt
langchain==0.0.171
openai==0.27.5
streamlit==1.23.1
feedparser
streamlit-auth0-component